server/swap: do not accept expired contracts in processInit
If the located contract transaction in processInit is found to have a
lock time that is in the past (already expired) revoke the match
immediately and respond to the user with an error.  No penalties are
assesed since processInit first verifies that the proper lock time has
been used.  This would be the case if the maker's swap took a very
long time to reach the required number of confirmations, and when
the taker's turn to publish their swap came around it was already
passed the expected lock time for their swap (matchTime + 8 hrs).

NOTE: client should be updated with similar logic to prevent attempting
to make expired contracts or contracts that are about to expire.