client/eth: Allow harness testing on testnet.
If testnet is set to true, create a new folder and wallets in the user's
home directory and change set up to create testnet nodes. Testing is
done pointing at testnet and the testnet contract.

Tests currently often fail because a transaction randomly cannot be
found to retrieve its receipt. This should not be an issue for swaps as
we only check for txn in tests to verify gas used. It may, however, have
implications for checking RegFeeConfirmations, as that similarly must
find a tx.