client/trade: Return counterPartyConfirms error.
Because the return data cannot all be trusted if the method has an
error, return that error up to the caller rather than just logging and
returning with possibly false values.

For swaps using contracts it is ok for the swap to be status none, in
which case we cannot check the lock time or confirmations on chain yet.
Do not log an error in this case.